# openapi.yaml
openapi: 3.1.0
info:
  title: Dejima API
  version: "0.1"
  x-api-version: 1
  description: >
    HTTP/WebSocket control surface for the Dejima daemon (`dejimad`). This is the
    same API the CLI/TUI use and the contract third-party apps build on.

    Status: Dejima is in alpha (0.x). The `v1/` path prefix is intended to be
    stable, but fields may change until `v1.0.0`; an integer `api_version`
    (advertised on `/v1/overview`) tracks the client/daemon contract separately
    from the semver tag. This spec is hand-maintained from the route table in
    `internal/api/server.go` and the request/response shapes in
    `internal/api/types.go` — those are the source of truth and the spec may
    drift slightly between releases.

    WebSocket endpoints (PTY session streams) are listed for discoverability but
    cannot be fully modeled in OpenAPI; see the per-path notes and the
    `SessionEnvelope` schema for the wire protocol.
servers:
  - url: http://{host}
    description: Daemon over the tailnet (DEJIMA_HOST, e.g. 100.84.12.7:7273)
    variables:
      host:
        default: 127.0.0.1:7273
security:
  - bearerToken: []
  - {}   # operator listeners (unix socket / tailnet) carry no per-request token
...
paths:
  /v1/islands:
    get:
      tags: [islands]
      summary: List all islands
      operationId: listIslands
      responses:
        "200":
          description: All islands, most-recently-used first
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/IslandInfo" }
    post:
      tags: [islands]
      summary: Create (provision) an island
      operationId: createIsland
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateIslandRequest" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CreateIslandResponse" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "409":
          description: An island with this name already exists
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
  /v1/islands/{name}:
    parameters:
      - $ref: "#/components/parameters/Name"
    get:
      tags: [islands]
      summary: Island detail (agents, presence, cpu/mem, git, health, disk, resources)
      operationId: getIsland
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/IslandInfo" }
        "404": { $ref: "#/components/responses/NotFound" }
    patch:
      tags: [islands]
      summary: Update cosmetic fields (title)
      operationId: updateIsland
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateIslandRequest" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/IslandInfo" }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      tags: [islands]
      summary: Purge an island (container + volumes)
      operationId: deleteIsland
      parameters:
        - name: force
          in: query
          schema: { type: boolean }
          description: Bypass the unpushed-work guard
      responses:
        "204": { description: Purged }
        "404": { $ref: "#/components/responses/NotFound" }
        "409":
          description: Refused — unpushed/dirty work (retry with force=true)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
...
  /v1/islands/{name}/events:
    parameters: [ { $ref: "#/components/parameters/Name" } ]
    get:
      tags: [islands]
      summary: The island's recent event log (newest first)
      operationId: islandEvents
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Event" }
        "404": { $ref: "#/components/responses/NotFound" }
...
  /v1/islands/{name}/exec:
    parameters: [ { $ref: "#/components/parameters/Name" } ]
    post:
      tags: [exec-and-files]
      summary: Run a one-shot command in an island
      operationId: exec
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ExecRequest" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ExecResponse" }
        "404": { $ref: "#/components/responses/NotFound" }
  /v1/islands/{name}/mailbox:
    parameters: [ { $ref: "#/components/parameters/Name" } ]
    get:
      tags: [mailbox]
      summary: Poll an island's intra-island agent mailbox
      operationId: pollMailbox
      parameters:
        - name: agent
          in: query
          schema: { type: string }
          description: Your agent id — filters to broadcasts plus messages addressed to you
        - name: since
          in: query
          schema: { type: integer, format: int64 }
          description: Only messages with seq greater than this cursor
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MailboxPollResponse" }
        "404": { $ref: "#/components/responses/NotFound" }
    post:
      tags: [mailbox]
      summary: Send a message into an island's mailbox (intra-island only)
      operationId: sendMailbox
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/MailboxSendRequest" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MailboxMessage" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "404": { $ref: "#/components/responses/NotFound" }
  /v1/islands/{name}/files/{path}:
    parameters:
      - $ref: "#/components/parameters/Name"
      - name: path
        in: path
        required: true
        schema: { type: string }
        description: File path within the island (defaults to the primary worktree)
    get:
      tags: [exec-and-files]
      summary: Read a file from an island
      operationId: readFile
      responses:
        "200":
          description: File contents
          content:
            application/octet-stream: { schema: { type: string, format: binary } }
        "404": { $ref: "#/components/responses/NotFound" }
    put:
      tags: [exec-and-files]
      summary: Write a file into an island
      operationId: writeFile
      requestBody:
        required: true
        content:
          application/octet-stream: { schema: { type: string, format: binary } }
      responses:
        "204": { description: Written }
        "404": { $ref: "#/components/responses/NotFound" }
  /v1/islands/{name}/logs:
    parameters:
      - $ref: "#/components/parameters/Name"
      - name: agent
        in: query
        schema: { type: string }
      - name: follow
        in: query
        schema: { type: boolean }
    get:
      tags: [exec-and-files]
      summary: Stream container/agent logs
      operationId: logs
      responses:
        "200":
          description: Log text
          content:
            text/plain: { schema: { type: string } }
  /v1/islands/{name}/port/scopes:
    parameters: [ { $ref: "#/components/parameters/Name" } ]
    get:
      tags: [port]
      summary: List Port grants (brokered host-folder scopes)
      operationId: listPortScopes
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PortScopesResponse" }
        "404": { $ref: "#/components/responses/NotFound" }
    post:
      tags: [port]
      summary: Grant a scoped host folder (deny-all by default)
      operationId: grantPortScope
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PortScopeRequest" }
      responses:
        "201":
          description: Granted
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PortScopeView" }
        "400": { $ref: "#/components/responses/BadRequest" }
  /v1/islands/{name}/port/scopes/{scope}:
    parameters:
      - $ref: "#/components/parameters/Name"
      - name: scope
        in: path
        required: true
        schema: { type: string }
        description: Scope name (e.g. scope-1)
    delete:
      tags: [port]
      summary: Revoke a Port grant by scope name
      operationId: revokePortScope
      responses:
        "204": { description: Revoked }
        "404": { $ref: "#/components/responses/NotFound" }
  /v1/islands/{name}/port/intake:
    parameters: [ { $ref: "#/components/parameters/Name" } ]
    post:
      tags: [port]
      summary: Brokered read-only file import (host -> island), ledgered
      operationId: portIntake
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PortIntakeRequest" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PortIntakeResponse" }
        "400": { $ref: "#/components/responses/BadRequest" }
  /v1/islands/{name}/port/export:
    parameters: [ { $ref: "#/components/parameters/Name" } ]
    post:
      tags: [port]
      summary: Brokered export (island -> host staging), append-only ledger
      operationId: portExport
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PortExportRequest" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PortExportResponse" }
        "400": { $ref: "#/components/responses/BadRequest" }
  /v1/islands/{name}/port/write:
    parameters: [ { $ref: "#/components/parameters/Name" } ]
    post:
      tags: [port]
      summary: Brokered write (island -> a read-write host scope), ledgered
      operationId: portWrite
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PortWriteRequest" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PortWriteResponse" }
        "400": { $ref: "#/components/responses/BadRequest" }
...
  /v1/capabilities/execute:
    post:
      tags: [capability]
      summary: Invoke a granted capability target (deny-all; every outcome ledgered)
      description: >
        Token-authenticated in-island callers are pinned to their own island
        (the `island` field is ignored); operators name the island explicitly.
      operationId: executeCapability
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CapabilityExecuteRequest" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CapabilityExecuteResponse" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "403":
          description: Capability not granted to this island
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "503":
          description: Capability execution unavailable on this host
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
...
  /v1/islands/{name}/link/send:
    parameters: [ { $ref: "#/components/parameters/Name" } ]
    post:
      tags: [link]
      summary: Send an info message to a specific agent in another island (granted channel)
      description: >
        The path island is the sender, pinned to the caller's token (it can only
        send as itself). Default-deny: with no matching operator grant the send is
        refused (403) and ledgered link.deny. A granted message is delivered into
        the target agent's ordinary mailbox with daemon-stamped cross-island
        Origin — so the response is a MailboxMessage.
      operationId: sendLink
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LinkSendRequest" }
      responses:
        "201":
          description: Delivered into the recipient's mailbox
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MailboxMessage" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "403":
          description: No link grant for this from→to/topic (cross-island is deny-all)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "404": { $ref: "#/components/responses/NotFound" }
...
  /v1/islands/{name}/link/action:
    parameters: [ { $ref: "#/components/parameters/Name" } ]
    post:
      tags: [link]
      summary: Request a named action on another island (deny-all; per-action approval)
      description: >
        The path island is the requester (token-pinned). Deny-all: requires a link
        grant for the channel AND the destination to expose the action. A
        pre-authorized action (in the grant's Actions allowlist) executes
        immediately (201, status "executed"); otherwise it is queued for operator
        approval (202, status "pending") and a link.action-pending event fires.
        Never a free-form prompt — only named, exposed action types.
      operationId: requestAction
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LinkActionRequest" }
      responses:
        "201":
          description: Executed (pre-authorized)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/LinkActionResponse" }
        "202":
          description: Queued for operator approval
          content:
            application/json:
              schema: { $ref: "#/components/schemas/LinkActionResponse" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "403":
          description: No grant, or the action is not exposed by the destination
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "404": { $ref: "#/components/responses/NotFound" }
...
  /v1/mcp/call:
    post:
      tags: [mcp]
      summary: Invoke a method on a granted MCP server (deny-all; every call ledgered)
      description: >
        Token-authenticated in-island callers are pinned to their own island
        (the `island` field is ignored); operators name the island explicitly.
        The brokered call is the JSON-RPC `method` (e.g. `tools/list`,
        `tools/call`) with opaque `params`.
      operationId: mcpCall
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/MCPCallRequest" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MCPCallResponse" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "403":
          description: MCP server not granted to this island
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
...
  /v1/tokens:
    get:
      tags: [tokens]
      summary: "List issued operator tokens (owner-only; never returns secrets)"
      operationId: listTokens
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TokensResponse" }
    post:
      tags: [tokens]
      summary: "Mint an operator bearer token (owner-only; secret returned once)"
      description: >
        Issues a role/island-scoped operator token. The bearer secret is returned
        exactly once in the response and is never stored in the clear — a lost
        secret means minting a new token, not recovering this one.
      operationId: createToken
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateTokenRequest" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CreateTokenResponse" }
        "400": { $ref: "#/components/responses/BadRequest" }
  /v1/tokens/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string }
        description: Token id
    delete:
      tags: [tokens]
      summary: "Revoke an operator token by id (owner-only)"
      operationId: revokeToken
      responses:
        "204": { description: Revoked }
        "400": { $ref: "#/components/responses/BadRequest" }
        "404": { $ref: "#/components/responses/NotFound" }
...
  /v1/internal/agent-event:
    post:
      tags: [daemon]
      summary: "Internal — agent telemetry sink (notify-hook shims). Not for app use."
      description: >
        Posted by the in-island notify hooks. A token-authenticated caller may
        only emit for its own island (the body's island is overridden by the
        token's island). Listed for completeness; third-party apps consume events
        via webhooks, not by posting here.
      operationId: agentEvent
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AgentEventRequest" }
      responses:
        "202": { description: Accepted }
        "400": { $ref: "#/components/responses/BadRequest" }
components:
  securitySchemes:
    bearerToken:
      type: http
      scheme: bearer
      description: Per-island bearer token (DEJIMA_TOKEN) for the autonomy path.
  parameters:
    Name:
      name: name
      in: path
      required: true
      schema: { type: string }
      description: Island name
...
  responses:
    NotFound:
      description: Island, agent, or resource not found
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
...
  schemas:
    ErrorResponse:
      type: object
      properties:
        error: { type: string }
    CreateIslandRequest:
      type: object
      required: [repo]
      properties:
        name: { type: string, description: optional; derived from repo if empty }
        repo: { type: string, description: git URL or local path }
        agent: { type: string, default: claude-code }
        image: { type: string, default: "dejima/island:latest" }
        cmd: { type: string, description: "entrypoint for agent=headless (required then; ignored otherwise)" }
        resources: { $ref: "#/components/schemas/Resources" }
        agents:
          type: array
          description: "Seed a multi-agent island; element 0 is the primary."
          items: { $ref: "#/components/schemas/AgentSpecRequest" }
        role: { type: string, description: "\"\" (work island) or \"home\" (assistant brain; must be headless)" }
        github_identity: { type: string, description: which daemon GitHub identity to clone/push as }
        seed_path: { type: string, description: host path bind-mounted read-only as the clone source (local daemon only) }
        owner: { type: string }
        tags: { type: object, additionalProperties: { type: string } }
    CreateIslandResponse:
      allOf:
        - $ref: "#/components/schemas/IslandInfo"
        - type: object
          properties:
            token:
              type: string
              description: only present on a token-authenticated (Home-Island spawn) create
    UpdateIslandRequest:
      type: object
      properties:
        title: { type: string }
...
    ExecRequest:
      type: object
      required: [cmd]
      properties:
        cmd: { type: array, items: { type: string } }
    ExecResponse:
      type: object
      properties:
        stdout: { type: string }
        stderr: { type: string }
        exit_code: { type: integer }
    MailboxSendRequest:
      type: object
      required: [payload]
      properties:
        from: { type: string, description: Sender agent id }
        to: { type: string, description: Recipient agent id; empty broadcasts to the island }
        topic: { type: string }
        payload: { type: string }
    MailboxMessage:
      type: object
      properties:
        seq: { type: integer, format: int64 }
        island: { type: string }
        from: { type: string }
        to: { type: string }
        topic: { type: string }
        payload: { type: string }
        time: { type: string, format: date-time }
        origin:
          $ref: "#/components/schemas/MailboxOrigin"
        action:
          $ref: "#/components/schemas/MailboxAction"
...
    MailboxPollResponse:
      type: object
      properties:
        messages: { type: array, items: { $ref: "#/components/schemas/MailboxMessage" } }
        latest: { type: integer, format: int64 }
...
    IslandInfo:
      type: object
      properties:
        name: { type: string }
        title: { type: string }
        repo: { type: string }
        agent: { type: string }
        image: { type: string }
        cmd: { type: string }
        role: { type: string }
        owner: { type: string }
        tags: { type: object, additionalProperties: { type: string } }
        state: { type: string, description: desired state from config }
        container: { type: string, description: observed runtime status }
        created_at: { type: string, format: date-time }
        last_used_at: { type: string, format: date-time }
        attached:
          type: array
          items: { $ref: "#/components/schemas/PresenceEntry" }
        stats: { $ref: "#/components/schemas/IslandStats" }
        agent_state: { $ref: "#/components/schemas/AgentStateInfo" }
        git: { $ref: "#/components/schemas/GitInfo" }
        health: { $ref: "#/components/schemas/IslandHealth" }
        disk: { $ref: "#/components/schemas/IslandDisk" }
        resources: { $ref: "#/components/schemas/Resources" }
        agents:
          type: array
          items: { $ref: "#/components/schemas/AgentInfo" }
...
    Event:
      type: object
      properties:
        type: { type: string, description: "e.g. island.created, agent.task-complete" }
        island: { type: string }
        agent: { type: string }
        timestamp: { type: string, format: date-time }
        payload: { type: object, additionalProperties: true }
    PortScopeRequest:
      type: object
      required: [host_path]
      properties:
        host_path: { type: string }
        mode: { type: string, enum: [ro, rw], default: ro }
    PortScopeView:
      type: object
      properties:
        name: { type: string }
        host_path: { type: string }
        mode: { type: string }
        granted_at: { type: string, format: date-time }
    PortScopesResponse:
      type: object
      properties:
        scopes:
          type: array
          items: { $ref: "#/components/schemas/PortScopeView" }
    PortIntakeRequest:
      type: object
      required: [scope, src_rel]
      properties:
        scope: { type: string, description: scope name to read from }
        src_rel: { type: string, description: path relative to the scope's host root }
        dest: { type: string, description: "container path; default /intake/<scope>/<src_rel>" }
    PortIntakeResponse:
      type: object
      properties:
        scope: { type: string }
        src: { type: string }
        dest: { type: string }
        bytes: { type: integer }
        sha256: { type: string }
...